refactor: UI shell 三个子系统按状态所有权归模块
按 4.5 节测量选块（零桥接、可搬状态最多）：

- app/sheets.js：sheet 开合原语（焦点管理/Tab 陷阱/导航层抬升）+ 通用确认弹窗，
  owns sheetFocusPrev/sheetKeyHandler/confirmResolve
  ★去掉一处认知泄漏：原 openSheet/closeSheet 直接读 permModal/questionModal 判阻断态、
    并伸手清审批模块的 arming 计时器；改为注入 blockingSheets()/onClosed()，通用原语不再认识具体弹窗
- app/drawer.js：左抽屉开合 + 边缘滑动手势，owns dragStartX/dragStartY/dragActive
- app/session-delete.js：两级删除会话，owns deleteTarget

app.js 7129→6960 行，顶层 let 90→85，app/ 模块 14→17。E2E 218 全过、单测 2289 全过。

同时记录：计划书原定的 3b「会话大区」经测量为最差候选（注入面 143、可搬状态仅 1、
需桥接 36），与已否决的 handler 表同属一类，不应照原样执行。
